Move FifoCompactionPicker to a separate .h file
