+++ db/compaction_picker_test.cc
@@ -4,10 +4,12 @@
 //  (found in the LICENSE.Apache file in the root directory).
 
 #include "db/compaction_picker.h"
+
 #include <limits>
 #include <string>
 #include <utility>
 #include "db/compaction.h"
+#include "db/compaction_picker_fifo.h"
 #include "db/compaction_picker_universal.h"
 
 #include "util/logging.h"
